requirements.txt
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
blinker==1.7.0
certifi==2024.2.2
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
dash==2.9.3
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.13
django-appconf==1.0.6
django-bootstrap-v5==1.0.11
django-browser-reload==1.12.1
django-extensions==3.2.3
django-import-export==3.3.7
django-node==4.0.2
django-tailwind==3.8.0
dpd_components==0.1.0
et-xmlfile==1.1.0
Flask==3.0.3
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
patsy==0.5.6
plotly==5.20.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
rcssmin==1.1.1
requests==2.32.2
rich==13.7.1
rjsmin==1.2.1
scipy==1.12.0
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
statsmodels==0.14.1
tablib==3.5.0
tenacity==8.2.3
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.2
xlrd==2.0.1
xlwt==1.3.0
